utils/mark-inputs/server/fw-mark-inputs/requirements.txt
-i https://pypi.org/simple
attrs==19.1.0
botocore==1.12.188
certifi==2019.6.16
chalice==1.9.1
chardet==3.0.4
click==6.7
docutils==0.14
enum-compat==0.0.2
flywheel-sdk==8.7.1
idna==2.8
jmespath==0.9.4
python-dateutil==2.8.0 ; python_version >= '2.7'
requests-toolbelt==0.9.1
requests==2.22.0
six==1.12.0
urllib3==1.25.3 ; python_version >= '3.4'
wheel==0.33.4
